Absolutify relative paths at build time :bug:
Previously, the base URL for all relative links was specified as part of
the generated HTML, in order to have links beginning with a `/` work
correctly in the local preview and not go to the filesystem root. This
worked but meant that links such as `#id`, which you would expect to go
to a heading on the current page, instead went to the homepage.

Now, a better implementation is used whereby relative links in the
output are directly replaced with the correct absolute link. This means
that the filesystem preview still works as before, but relative links
created by the user behave as they would normally do.

Fixes #1.