Process template tags within Nix :zap:
This removes the need to call out to an external Rust program
every time a template tag is encountered, therefore speeding
up evaluation. Furthermore, there is now no need to download a
Rust compiler and libraries the first time Coricamu is used on
a machine.

The code is currently quite disorganised. I originally intended
to match both the opening and closing tag as a single pattern, with
a capturing group for the contents between them. However, the type
of regular expressions which Nix supports doesn't include lazy
multipliers. This meant that if the same template was used multiple
times, we would match all of the stuff between the two template calls
as the contents of a single template, rather than correctly detecting
them as two separate ones.

Nix could do with switching to PCRE2 regular expressions to support
more exotic use cases such as this. I've saved the original code I
developed as a patch file, locally, so that it can be brought back if
this is ever done.

Anyway, there are still some bugs in this code:

- You can't nest a template as the contents of itself, because this
  case isn't currently handled.
- Invalid opening and closing of template tags also isn't handled.
  (For example, not closing a tag, or closing a tag when none was
  opened.) This should raise an error of some sort.