Begin rewriting in functional style :boom: :construction:
This is intended to encourage more organised code, by only allowing a
larger component (a website) to import smaller components within it
(pages), without allowing definitions for the same component to be
spread across multiple files.

Component functions will be extended with `makeOverridable` so that
their results can be modified by parent components, for example the
function for a website will override pages passed to it such that they
contain the site-wide header.

This rewrite will also include a rework of the templating system, and
improvements to the handling of images, styles and scripts.