Improve handling of auxiliary files :recycle:
The new subdirectory function can be used to move a page and its
auxiliary files into a subdirectory.

It moves files unless their name begins with a slash. So an auxiliary
file named "a" is kept alongside its page, whereas an auxiliary file
named "/a" is lifted to the root of the site.

Creating a page named "a/b" would place its auxiliary files at the "a"
level, whereas creating a page named "b" and then passing it into the
subdirectory function would place the auxiliary files at the same level
as "b".

Subdirectories can also pass a `makeManifests` function which will be
called with all of the files in the current directory. This will be used
to generate sitemaps, RSS feeds and so on.