Add functions for project documentation :sparkles:
These generate a standard documentation site for a GitHub project,
reducing boilerplate.

Any markdown files with names in all caps will automatically be added to
the site and linked from the header. We try to extract a page title from
markdown but will fall back to the file name if no title is found.